Use apt-helper to get contents of Sources/Packages files
As described in <20160111171230.GA17291@debian.org>, the files under
/var/lib/apt/lists may be compressed (due to Acquire::gzipIndexes).
Instead of directly parsing the files, which will fail if they're
compressed, try to use /usr/lib/apt/apt-helper's cat-file command to get
the contents.

When that isn't available, fallback to the old method of trying to
directly parsing the files.

Signed-off-by: James McCoy <jamessan@debian.org>